requirements.txt
-i https://pypi.org/simple
alembic==1.1.0
certifi==2019.6.16
chardet==3.0.4
click==7.0
decorator==4.4.0
flask-migrate==2.5.2
flask-sqlalchemy==2.4.0
flask==1.1.1
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
mako==1.1.0
markupsafe==1.1.1
paho-mqtt==1.4.0
pbr==5.4.2
psycopg2-binary==2.8.3
psycopg2==2.8.3
python-dateutil==2.8.0
python-editor==1.0.4
pyyaml==5.1.2
requests==2.22.0
six==1.12.0
sqlalchemy-migrate==0.12.0
sqlalchemy==1.3.8
sqlparse==0.3.0
tempita==0.5.2
urllib3==1.25.3
werkzeug==0.15.5
